backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.3.0
click==8.3.1
colorama==0.4.6
cryptography==45.0.2
ecdsa==0.19.1
fastapi==0.135.3
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.4
pandas==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlalchemy==2.0.41
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.42.0
